relocated handlers in one global handler  and deleted separeted handlers / fixed css rule for login button
