+++ sass/components/login/_login.scss
@@ -4,7 +4,7 @@
 
 .form__input__wrapper_login {
   &:last-child {
-    margin-bottom: 16px;
+    margin-bottom: 64px;
   }
 
   &:nth-child(1)::before {
